Finish moving logic to the backend
