+++ pyproject.toml
@@ -108,6 +108,7 @@ module = [
     'ruamel.*',
     'scipy.*',
     'sqlalchemy.*',
+    'sqlalchemy_utils.*',
     'tqdm.*',
     'wrapt.*',
 ]
